move Paginator into state store package
